src/requirements.txt
Flask==2.2.2
SQLAlchemy==1.4.40
gunicorn==20.1.0
paramiko==2.12.0
PyMySQL==1.0.2
coloredlogs==15.0.1
python-dotenv==0.20.0
